Add Edit and Delete Buttons to Book Component
Conditionally display edit and delete buttons on the `Book`
component if the user is signed in and the current user is the
owner of the book. Display a message if the user is not signed
in or if the signed in user does not own the book. Add the
user as a prop to the Book component in the route in `App.js`.
Display the originalLanguage and firstPublished values in the
`Book` component.  Format the `firstPublished` date to display
the date in a user-friendly manner.