Replace Bootstrap .css file with .scss file
The react-auth-template currently imports the compiled and
minified bootstrap.css file in `index.js`. In order to be able
to use the sass features in Bootstrap, removed this import and
added the scss file import into the `src/css/index.scss`
file.  Also, you removed all of the styles in `src/
App.scss` and `src/css/index.scss`.